Avoid the activity is shown incorrectly
When switching user quickly, the user is not stopped and switch back, the last activity will be resumed. If the user state is changed from STOPPING to RUNNING_UNLOCKED, there is no problem, but when the user state is changed from SHUTDOWN to BOOTING, there will be the problem, some activities without directBootAware will be launched before user is unlocked and then cause crash. So add user stated check and activty directBootAware check.

Bug: 201568631
Test: switch quickly back and forth between the two users and check if there is any crash. atest ActivityRecordTests pass, no impact

Change-Id: I40d554f1fab047deb3acc643c6cac7d9f6e6c7bb
Signed-off-by: GhostMaster69-dev <rathore6375@gmail.com>