Fix the bug in verifyApp method that makes it completely useless
Renew verifyApp method to throw required PackageManager.NameNotFoundException
when the provided package doesn't exist on the device. The current
implementation of this method could not throw the required exception.
This leads to the fact that any packages, including non-existent ones,
are able to get into the vpnconfig (allowedApplications/disallowedApplications).
Such a corrupted config can cause VPN applications to crash/destabilize.

Test: manual test app
Change-Id: I95333dba5f25eb17741aefc6dbd265d541272116
Signed-off-by: GhostMaster69-dev <rathore6375@gmail.com>