Not run hideDotView as an animation
When click “Video” button in camera view, the mediaRecorder will run prepare(). With the same time, screenDecoration in statusbar will run hideDotView() as an animation.

It will make “mediaRecorder run prepare()” cost more time than S load.
As in android S, the animation of hideDotView() will not run.
So suggest not to run hideDotView() as an animation, but run function directly.

Bug: 240375342
Test:Manual
1. Open camera app.
2. Let the camera pattern be Picture.
3. Click the “Video” button in camera view, let the pattern change from Picture to Video.

Change-Id: I76f310fd96a1c07c1c3739c7d671b00a9301a19f
Signed-off-by: GhostMaster69-dev <rathore6375@gmail.com>