Fix LoadedApk resource dir isn't up to date issue
When a context is created through Context#createApplicationContext,
the underlying LoadedApk cached in ActivityThread#mResourcePackages
is updated using the ApplicationInfo passed in. It creates unexpected
problems when using APIs that expect to be able to read resources from
a previous version of a package after that package as been updated.

We use ActivityThread#isLoadedApkResourceDirsUpToDate to determine if
we should update ApplicationInfo of LoadedApk. But it is not reliable,
for ApplicationInfo#resourceDirs is subset of ApplicationInfo#overlayPaths.

With this change, we use ApplicationInfo#sourceDir and publicSourceDir.

Test: atest FrameworksCoreTests:ContextTest

Signed-off-by: liulvping <liulvping@xiaomi.com>
Change-Id: Ia787d33067c7a7218ef0765840d5d39187ad9755
Signed-off-by: GhostMaster69-dev <rathore6375@gmail.com>